Move all benchmarks into a separate directory
