+++ setup.cfg
@@ -10,5 +10,5 @@ inline-quotes = "
 
 [tool:pytest]
 minversion = 3.0
-norecursedirs = build c docs temp .cache .git .idea .venv
+norecursedirs = build c docs temp benchmarks .cache .git .idea .venv
 python_files = test_*.py bench_*.py
